Add working option menu and popup menu
In this commit, an option menu is added, in which you can set whether
you automatically sign up for both normal and resits. These preferences
can also be changed for each course individually. Mroeover, a popup menu
is created with which you can update the courses and the exams for the
discovered courses.