Add simple debug functionality
In this commit, an easy way to enable debug functionality is added. This
debug functionality includes buttons to clear the sync, the courses or
the exams respectively. The debug mode is enabled by a boolean in
options.js, which will display these buttons underneath the title in the
options page when true.