Add error handling for Osiris errors
In this commit, basic error handling for errors posted by Osiris is
applied. This means that every time that Osiris throws an error, a popup
is created with the error message, stating that this has to be reported.

Three cases of errors commonly occurring are removed:
- Registering for an exam for which you are already registered (by
changing the updating sequence).
- Selecting nothing for (de)registering (by checking this in the content
scripts)