Add continuous integration and change structure
In this commit, a basic package.json is added to the project, which
specifies ESLint as a dependency, as well as a configuration for ESLint.
Moreover, the structure of the project is changed to have everything
that is written as source being in the src/ folder. The packages file is
excluded, as this contains third party packages which are used in the
application (and which have to be shipped with the extension)